Allow extending enumerators

+++ climmob/views/enumerator.py
@@ -13,20 +13,19 @@ from climmob.processes import (
     getEnumeratorData,
 )
 from climmob.views.classes import privateView
+import climmob.plugins as p
 
 
 class getEnumeratorDetails_view(privateView):
     def processView(self):
-
         if self.request.method == "GET":
-
             userOwner = self.request.matchdict["user"]
             enumId = self.request.matchdict["enumid"]
             enumerator = getEnumeratorData(userOwner, enumId, self.request)
             self.returnRawViewResult = True
-
+            for plugin in p.PluginImplementations(p.IEnumerator):
+                enumerator = plugin.before_returning_context(self.request, enumerator)
             return enumerator
-
         raise HTTPNotFound
 
 
@@ -49,19 +48,30 @@ class enumerators_view(privateView):
                 if not enumeratorExists(
                     self.user.login, dataworking["enum_id"], self.request
                 ):
-                    added, message = addEnumerator(
-                        self.user.login, dataworking, self.request
-                    )
-                    if not added:
-                        error_summary = {"dberror": message}
-                    else:
-                        dataworking = {}
-                        self.request.session.flash(
-                            self._("The field agent was created successfully.")
+                    continue_add = True
+                    message = ""
+                    for plugin in p.PluginImplementations(p.IEnumerator):
+                        if continue_add:
+                            continue_add, message = plugin.before_adding_enumerator(
+                                self.request, self.user.login, dataworking
+                            )
+                    if continue_add:
+                        added, message = addEnumerator(
+                            self.user.login, dataworking, self.request
                         )
-                        # if nextPage:
-                        #    self.returnRawViewResult = True
-                        #    return HTTPFound(nextPage)
+                        if not added:
+                            error_summary = {"dberror": message}
+                        else:
+                            for plugin in p.PluginImplementations(p.IEnumerator):
+                                plugin.after_adding_enumerator(
+                                    self.request, self.user.login, dataworking
+                                )
+                            dataworking = {}
+                            self.request.session.flash(
+                                self._("The field agent was created successfully.")
+                            )
+                    else:
+                        error_summary = {"dberror": message}
                 else:
                     error_summary = {
                         "exists": self._("This field agent username already exists.")
@@ -82,21 +92,37 @@ class enumerators_view(privateView):
                 else:
                     dataworking["enum_active"] = 0
 
-                mdf, message = modifyEnumerator(
-                    self.user.login, enumeratorid, dataworking, self.request
-                )
+                continue_update = True
+                message = ""
+                for plugin in p.PluginImplementations(p.IEnumerator):
+                    if continue_update:
+                        continue_update, message = plugin.before_updating_enumerator(
+                            self.request, self.user.login, enumeratorid, dataworking
+                        )
+                if continue_update:
+                    mdf, message = modifyEnumerator(
+                        self.user.login, enumeratorid, dataworking, self.request
+                    )
+                    if not mdf:
+                        error_summary = {"dberror": message}
+                        dataworking["enum_password"] = decodeData(
+                            self.request, dataworking["enum_password"]
+                        ).decode("utf-8")
 
-                if not mdf:
+                    else:
+                        for plugin in p.PluginImplementations(p.IEnumerator):
+                            plugin.after_updating_enumerator(
+                                self.request, self.user.login, enumeratorid, dataworking
+                            )
+                        dataworking = {}
+                        self.request.session.flash(
+                            self._("The field agent was modified successfully.")
+                        )
+                else:
                     error_summary = {"dberror": message}
                     dataworking["enum_password"] = decodeData(
                         self.request, dataworking["enum_password"]
                     ).decode("utf-8")
-
-                else:
-                    dataworking = {}
-                    self.request.session.flash(
-                        self._("The field agent was modified successfully.")
-                    )
 
         return {
             "activeUser": self.user,
@@ -115,15 +141,30 @@ class deleteEnumerator_view(privateView):
         enumeratorid = self.request.matchdict["enumeratorid"]
 
         if self.request.method == "POST":
-            deleted, message = deleteEnumerator(
-                self.user.login, enumeratorid, self.request
-            )
-            if not deleted:
+            continue_delete = True
+            message = ""
+            for plugin in p.PluginImplementations(p.IEnumerator):
+                if continue_delete:
+                    continue_delete, message = plugin.before_deleting_enumerator(
+                        self.request, self.user.login, enumeratorid
+                    )
+            if continue_delete:
+                deleted, message = deleteEnumerator(
+                    self.user.login, enumeratorid, self.request
+                )
+                if not deleted:
+                    self.returnRawViewResult = True
+                    return {"status": 400, "error": message}
+                else:
+                    for plugin in p.PluginImplementations(p.IEnumerator):
+                        plugin.after_deleting_enumerator(
+                            self.request, self.user.login, enumeratorid
+                        )
+                    self.request.session.flash(
+                        self._("The field agent was successfully removed")
+                    )
+                    self.returnRawViewResult = True
+                    return {"status": 200}
+            else:
                 self.returnRawViewResult = True
                 return {"status": 400, "error": message}
-            else:
-                self.request.session.flash(
-                    self._("The field agent was successfully removed")
-                )
-                self.returnRawViewResult = True
-                return {"status": 200}
